refactor mapping logic to `FeedMapper` class
